C2_Server/requirements.txt
Flask
gunicorn
werkzeug
Flask-SocketIO
Flask-Limiter
psycopg2-binary